Use concrete types for `OsStr::display`
